requirements.txt
manim==0.20.1
pydub>=0.25.1
edge-tts>=6.1.0
pyyaml>=6.0
svgwrite>=1.4.3
